Move rollback.go to rollback package inside reconcilers directory
