+++ pkg/discoveryservice/manager.go
@@ -12,6 +12,7 @@ import (
 	marin3rv1alpha1 "github.com/3scale/marin3r/apis/marin3r/v1alpha1"
 	marin3rcontroller "github.com/3scale/marin3r/controllers/marin3r"
 	envoy "github.com/3scale/marin3r/pkg/envoy"
+	rollback "github.com/3scale/marin3r/pkg/reconcilers/marin3r/envoyconfig/rollback"
 	"github.com/go-logr/logr"
 	"k8s.io/apimachinery/pkg/runtime"
 	util_runtime "k8s.io/apimachinery/pkg/util/runtime"
@@ -90,7 +91,7 @@ func (dsm *Manager) Start(ctx context.Context) {
 			ClientAuth:   tls.RequireAndVerifyClientCert,
 			ClientCAs:    loadCA(dsm.CACertificatePath, setupLog),
 		},
-		marin3rcontroller.OnError(dsm.Cfg),
+		rollback.OnError(dsm.Cfg),
 		setupLog,
 	)
