+++ pkg/reconcilers/marin3r/envoyconfig/rollback/on_error.go
@@ -1,4 +1,4 @@
-package controllers
+package rollback
 
 import (
 	"context"
